Rename anomolies.js to autoscan.js
