+++ src/server/autoscan.js
